+++ app/rayshift/quest.py
@@ -57,16 +57,25 @@ async def get_quest_response(
 
 
 async def get_quest_detail(
-    conn: AsyncConnection, region: Region, quest_id: int, phase: int
+    conn: AsyncConnection,
+    region: Region,
+    quest_id: int,
+    phase: int,
+    questSelect: int | None = None,
 ) -> Optional[QuestDetail]:
-    db_quest_detail = await get_rayshift_quest_db(conn, quest_id, phase)
+    db_quest_detail = await get_rayshift_quest_db(conn, quest_id, phase, questSelect)
     if db_quest_detail:
         return db_quest_detail
     else:
         quest_response = await get_quest_response(region, quest_id, phase)
         if quest_response and quest_response.questDetails:
             await insert_rayshift_quest_db(conn, quest_response.questDetails)
-            return next(iter(quest_response.questDetails.values()))
+            quest_detail = next(iter(quest_response.questDetails.values()))
+            if (
+                questSelect is not None and quest_detail.questSelect != questSelect
+            ):  # pragma: no cover
+                return None
+            return quest_detail
 
         return None