+++ tests/test_nice.py
@@ -72,6 +72,8 @@ test_cases_dict: dict[str, tuple[str, str]] = {
     "quest_NA_enemy_cache": ("NA/quest/94034503/1", "NA_CCC_Detour_3"),
     "quest_JP_remapped": ("JP/quest/94060012/1", "JP_remapped_quest"),
     "quest_JP_support_servant": ("JP/quest/94051406/1", "JP_support_servant"),
+    "quest_NA_select_0": ("NA/quest/3000109/1", "NA_quest_LB1_select_0"),
+    "quest_NA_select_1": ("NA/quest/3000110/1", "NA_quest_LB1_select_1"),
     "ai_beni_cq_monkey_NA": ("NA/ai/svt/94032580", "NA_AI_Beni_CQ_monkey"),
     "kh_cq_JP": ("JP/ai/field/90161870", "JP_KH_CQ_taunt"),
     "bgm_NA_with_shop": ("JP/bgm/138?lang=en", "JP_BGM_Shinjuku"),