+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to the public API will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Changes to be tagged
+
+## 5.76.0 - 2022-01-23
 ### Added
+- support for `questSelect` quests.
 - `afterClear` and `spotName` to basic quest and basic quest phase.
 - `recommendLv` and `spotName` to nice quest and nice quest phase.
 - `call` to nice stage.